renamed and moved styles.js to styles
